Moved imsim/phosim specialisations into their own files/directories
I had originally thought that imsim, phosim, and the real lsstCam
would all be the same.  This isn't true for a number of reasons,
so this cleanup is desireable.  In particular, it allows us to use
a different ParseTask for each variant, and this in turn (now)
allows us to use a different camera (e.g. with/without corner rafts)

The phosim mapper has a couple of hacks in it that should be fixed
upstream:
   OBSID is sometimes a string, not an int
   FILTER is a random string for the wavefront chips