Switch to using just the visit number instead of (dayObs, seqNum) tuples
When ingesting seqNum change to testSeqNum to ensure this is not confused
with seqNum in auxTel which has a very different meaning.

Remove use of seqNum in calculating the visit number.

Change the templates in the ts8 and auxTel mapper (which ts8 inherits from).
This included tidying up so that keys were always defined in the same order
and removing redefinitions of things that were being inherited already.

Remove detectorName from auxTel policy as there is no need for it.

Add support for ingesting multiple different rafts by including their serials.