Add configs for visitAnalysis and coaddAnalysis
This provides the config overrides required by the visitAnalysis and
coaddAnalysis scripts in the pipe_analysis repo for the reference object
loader in obs_lsst, including loading the filter map.

The various config parameters that are set to False here can/should be
removed under the following circumstances:

doApplyUberCal=False -> remove when uber calibration (i.e. jointcal) is
                        being regularly run

doWriteParquetTables=False -> remove is persisted parquet tables for
                              interactive analysis of lsst-dm/qa_explorer
                              are desired

doApplyColorTerms=False -> remove if color terms for reference catalog to
                           camera filters are added and need to be applied
                           for the reference catalog comparisons