Add code to generate defect files and registry
Add a SConscript file which calls genDefectFits.py
and genDefectRegistry.py to build fits files from the
defect.dat files, and create a defect registry at
scons-time.

This is following the pattern (and copying the code)
in obs_subaru.

It is deliberately not moving this code into obs_base
nor implementing this for other parts of obs_lsst
as this is a pattern we want to change, but there is no
point in doing this before Gen3 or fixing up calibration
stuff in general.