Adopt LSST boresightRotAngle convention for imsim
The LSST convention for the boresight rotation angle is such that,

At boresightRotAngle = 0, the focal plane pixels are alligned:
+x: E->W (-ve RA), +y: S->N (+ve Dec)

To match this convention, the imsim data requires the following:

boresightRotAngle = 90 deg - ROTANGLE

where ROTANGLE is the FITS header card in the raw imsim data.

Prior to the "translator" method of standardizing metadata from obs
packages to the consistent set of LSST conventions, this rotation
was being accommodated in lsstCameraMapper.py.  This removes the
latter accommodation and adds it to the imsim translator.