Change calibration refCols and reference table.
refCols is the list of data id keys that may be specified (their
combination must select a unique dateObs). 'visit' is retained for a
future time when 'snap' can also be included.

The reference table was attempting to use a visit optimization, but ISR
will in general act on exposures, not visits, so we change to the 'raw'
table.