Update R22 for future values
The master version of the camera has ITL sensors installed in
R22, but that's not what ended up in the manufactured camera.

BOT data from R22 were used in the assembly test and
so this branch needs to implement the as built version for at
least that raft.

All other rafts and gain/saturation/read noise for this raft are
implemented in DM-22605 and will land when that lands.