Add template overrides to policy files for imsim
Prior to DM-23584, the templates in obs_lsst/policy/lsstCamMapper.yaml were
compatible with imsim conventions.  However, the change in the expId length
from 8 to 13 digits in DM-23584 rendered them incompatible with the imsim
standards (which remains "08d" for the expId field).  This adds the policy
file policy/imsim/imsimMapper.yaml with the appropriate overrides and
updates python/lsst/obs/lsst/imsim.py to apply them.