Fix lsstCamMapper.yaml levels and FFP templates
Issues were noticed when trying to run showCamera, which produced a seperate
file for each detector (rather than one for the entire focal plane).  This was
also true for pipe_drivers' visualizeVisit.py whose purpose is to create a
full focal plane view of a given visit in a single fits file.  As such, these
templates should not be subdivided further than visit (i.e. omit divisions and
any references to raftName, detectorName, and detector).  The above remedies
the "one file created per detector issue", but there was still an issue with
the fact that the levels were not set, such that a full focal plane file was
created, but with data from only a single detector in it (and there was a loop
over each detector with the file overwritten with each iteration showing the
single detector data at each step).  The levels, their defaults, and, in
particular, the specification of the dataId keys that are NOT relevant to a
particular level, are updated here.