Delete comCamMapper.yaml to remove redundancies
The comCamMapper.yaml file was (almost) entirely redundant with the templates
already (and indentically) defined in the default lsstCamMapper.yaml.  The few
exceptions were instances of %(expId)d in the former that are actually
%(expId)013d in the latter (e.g. for dcrDiff_diaSrc).  I do believe there was
an oversight in the 08d -> 13d conversion for the expId length of commit:
https://github.com/lsst/obs_lsst/commit/b2b3add2b45b5959b87e50cb79c42e102e0b4c1f
such that they are indeed supposed to have indentical templates and hence the
removal of the comCamMapper.yaml override file is the correct thing to do.