Properly remove local camera cache
In DM-25923 I removed half the cache but that broke caching
completely so we ended up with tests failing because
LSSTCam would cache and then the other subclasses would
think that they didn't need to cache because the parent
class had the camera.

Whether tests failed depended on how many cores were in use
for testing and the order that pytest ran tests.

Now fully remove the vestigial caching and completely rely
on yamlCamera caching.