Add more ND filters and use base part of filter in coaddExposureId
With all the ND filters in the second BOT filter wheel we have more than 128 filters, so would need an extra bit in nbit_filter. Instead, we only use the "base" of the filter name (discarding anything after the ~) when calculating coaddExposureId.  This is OK because we won't ever be coadding data of this type;  additionally we assert that the ID of any filter that appears in a coaddExposureId fits in the allocated bits, and arrange that the real LSST filters satisfy this condition.