Adapt to gen2to3 conversion and tests being stricter.
Base defect tests now fail when there are no defects; this was not
something I changed intentionally, but I think silently not failing
(as before) was not terribly useful behavior for a test for other
instruments.

In addition the converter itself now complains where there are datasets
in a root repository with no run configured for them (like the
processCcd_metadata dataset in the PhoSim repo here), instead of
silently ignoring them.