Interpret detector_max as exclusive bound on zero-based detector_nums.
The naming isn't great, but this is the interpretation that daf_butler
and other cameras have assumed, and we need obs_lsst to be the same
in order for DimensionPacker to behave consistently with
astro_metadata_translator until we can unify that log.

This removes the DETECTOR_MAX override for TS3 because it should just
inherit the same value from its base class anyway.