Remove concat option from compute_detector_exposure_id implementations.
This makes it harder for these to get out of sync with
daf.butler.DimensionPacker implementations.

It also requires that we change the DETECTOR_MAX values for TS8 and
UCDCam to powers of ten, and that means we can't easily check in the
ID calculation that the detector values given are physically valid, if
they lie between the true value and the power of ten above that (hence
the removed test on this commit).  We could of course introduce a new
translator field to hold that number, but it would be nice if we could
actually call that number DETECTOR_MAX, all other things being equal,
and we can't do that (at least not without creating extra confusion)
without changing the daf_butler dimension schema, and we can't do that
without database migrations.  So I think this is the best
local-minimum best solution for now, and we should revisit this in the
future when we're willing to change more, whether that's the
daf_butler schema, the actual ID calculations for various cameras, or
something else.