Fix raw.detector implementation to match raw exposure reading.
Support for raw.detector is new in obs_base, but requires work for LSST
because we modify the detectors when raw images are read.  But it's
also more important for LSST, because this gives us a way to get a
detector with the right geometry for raw, without actually reading the
image.