Actually read and propagate raw amp metadata to assembly code.
Downstream functions like fixAmpsAndAssemble were expecting per-amp
header metadata to be passed in, but were getting empty PropertyLists
because we read the data from disk via Image, not Exposure.  We never
noticed because they were designed to do nothing when those header keys
were missing.

This did not affect the metadata ultimately associated with raws (at
least in Gen3) because we read that separately.