Disable configs and columns about bright object mask flags.
DM-30649 fixed a config option in assembleCoadd that makes it so
the BRIGHT_OBJECT mask plane does not exist, while before it was just
never being set (because we don't have any bright object masks for
non-HSC instruments, and we never will, at least not in this form).

But this broke downstream processing that expected this mask plane to
exist.