Turn on delta mag rejection in astrom for imsim
This turns on the functionality (added on DM-30490) in the SFM astrometric
calibration stage to perform a rejection on delta mag (ref - src) outliers
in the match catalog before sending it to the WCS fitter.  Fairly extensive
testing on LSSTCam-imSim data has proven this to be a most desirable option
in that it fixes (i.e. a good astrometric fit is now found) almost all of
the bad fit cases and it does-no-harm for the cases that already had good
fits (see DM-30812 for details of this validation).