merge extra hard-coded named headers in rawFormatter
In current LATISS data (and likely future comcam & LSSTCAM data), the FITS files may contain extra HDUs "REB_COND" and "CONFIG_COND", which describe some sensor state that does not make sense to be included per-amplifier. This change checks if those HDUs are present, and if they are includes them in the metadata.