Ensure positive/non-zero detector_max in _instrument.py
The LATISS camera has DETECTOR_MAX set to 0 (to ensure the
desired behaviou when calculating the detector_exposure_id).
However, this value is also used in _instrument.py to set
"detector_max" which is used by packers to set data ids.
Having this set to 0 resulted in the "maxBits" used to in the
data id "packers" to always return 0 for the "visit_detector"
id.  This casued issues (at least) with the makeCcdVisitTable
task in pipe_tasks as it uses this id as the *unique* index
value for a pandas dataframe.  The solution here is to ensure
that detector_max gets a minimum value of 1.