Update astrometry configs for LATISS
Now that the pointing accruacy and stability has improved for AuxTel,
the bbox padding for reference catalog loading can be greatly reduced.
This has the benefit of avoiding many very bad fits where the matcher
locked onto sources way outside of the true field for the exposure.

Note that this update is likely only valid for data taken as of 2022-05
(when the main pointing improvements occurred).  If processing earlier
data, the configs may need to be relaxed to their previous values to
avoid astrometric failures.

This also turns on the reference vs. source magnitude outlier rejection
to help in avoiding bad matches.