Use objectSize source selector for LATISS
With a few config override settings, the objectSize selector works well
with LATISS data (and these settings recovered half, 12 of 24,  of the
previously failing PSF model attempts in a test dataset).

Some minor updates to the astrometric matcher are also made to help
with cases previously failing to find matches.