Fix TS8 exposure id calculation for historical data
We have already ingested TS8 data into a butler using the old
timestamp exposure_id calculation. It is very disruptive to change
this because it will require the entire repository be remade from
scratch. Furthermore, the fix to DATE-OBS broke all the historical
calculations where the exposure_id was being derived from the end
of the integration by mistake.

This commit retains LSSTCam exposure_id for > 2023-05-24 and
uses the original DATE-OBS value to calculate the exposure ID for
older data.