Restore the original max exposure ID for TS8
This restores the TS8 exposure id maximum to a value before
we switched over to using LSSTCam exposure_id. We need this
to take into account the earlier data that use a timestamp
and so a larger exposure ID. We do not change the value since
we do not want to modify existing repositories.