Set mag limits for refcat loading in LSSTComCamSim
These have been selected to more than span the expected magnitude
range for the icSrc catalog sources (those available for the
calibrations).  Loading in reference sources outside of this range
is undesirable for a few reasons, including wasting time on objects
that are not acutally in contention for matcing and, more importantly,
leaving a danger of the fit getting "lost" by locking onto the
"superfluous" reference catalog objects (we saw this happen in
photoCal for 3 of 25935 dataIds in the first round Ops Rehearsal 3
dataset).