Derive AZEL from RA/DEC for simulated comcam
This is currently problematic because the simulated files
use a DATE-BEG that differs from the date used to calculate
the airmass and hour angle. It's not clear what the right
thing to do is for azel calculation since calculating
the actual values for the DATE-BEG might end up with
a negative elevation. Trying to calculate the azimuth from the
hour angle / declination / airmass would lead to an azimuth
that is not consistent with the time and ra/dec.