Update Sim test files so they pass tests
Neither native ComCamSim or LSSTCamSim files use the correct
definition of observing day. This meant that now that we are
testing for conformance they failed the tests.

Use a ComCamSim file from OR3 that was generated
from the test stand (and so does conform). This file was also
zeroed out to take up less space.

There are no conformant LSSTCamSim files at this time so the
file was rewritten with an MJD-OBS that does match the
day obs.

It is not posible to use the controller code when calculating
the observing day offset.