Add per-band spatial order for piff override
We now have the option to set a per-band config for the spatial order
of the final PSF modle fit.  This sets our preferred overrides for
LSSTComCam, which are motivated by the fact that u band will almost
always have fewer PSF canditates than the other bands (throughput,
stellar SED, etc.), and that u-band coadds are highly unlikely to be
used in lensing/shear analyses. Setting a smaller spatial order for
u precludes an attempt at a higher order fit that inevitably won't be
well-constrained (and currently piff does not handle this internally
and returns a dodgy PSF model).