Update minNPsfStarPerBand for LSSTComCam
The default spatial order for the piff-based PSF model is modified
for LSSTComCam processing.  This updates the mininum number of PSF
stars used in the fit required to be included in the coadd to match
the per-band spatial order settings for LSSTComCam (with the fallback
value matching the task default).  Note that piff now does this check
internally, so this should be a no-op as long as piff is used for the
final PSF model (but is kept as a safeguard).