Add overrides for coadd detection
In particualr, we have found that the "scaleVariance", "thresholdScaling",
and "backgroundTweak" detection adaptations have been proven to have
potential to do more harm than is reasonable for any potential improvements,
thus we set them all to False here. The root causes leading to their addition
(for HSC data) should be assessed and, if present, fixed instead.