Re-increase astrometry pixel margins for LSSTCam
These were tightened on a misplaced belief that the pointing model for
LSSTCam had been significantly improved as of an adjustment made on
April 25, 2025.  This turns out not to be the case and we continue to
get very large pointing offsets, so these large margins are still
required for LSSTCam data processing.